Refactor generator.dict
