+++ lib/dict.js
@@ -7,22 +7,6 @@ var pair = require("./pair.js");
 var string = require("./string.js");
 var utils = require("./utils.js");
 
-function fromArray(arrayOfPairs) {
-  var res = {};
-  arrayOfPairs.forEach(function (p) {
-    res[p[0]] = p[1];
-  });
-  return res;
-}
-
-function toArray(m) {
-  var res = [];
-  Object.keys(m).forEach(function (k) {
-    res.push([k, m[k]]);
-  });
-  return res;
-}
-
 function makeMapShow(elShow) {
   return function (m) {
     return "{" + Object.keys(m).map(function (k) {
@@ -38,7 +22,7 @@ function dict(arb) {
   var pairArbitrary = pair.pair(string.string, arb);
   var arrayArbitrary = array.array(pairArbitrary);
 
-  return arrayArbitrary.smap(fromArray, toArray, makeMapShow(arb.show));
+  return arrayArbitrary.smap(utils.pairArrayToDict, utils.dictToPairArray, makeMapShow(arb.show));
 }
 
 module.exports = {
