+++ lib/generator.js
@@ -225,18 +225,17 @@ function generateNEArray(gen) {
   return utils.curried2(result, arguments);
 }
 
-// TODO: remove
 /**
-  - `generator.map(gen: generator a): generator (map a)`
+  - `generator.dict(gen: generator a): generator (dict a)`
 */
 function generateDict(gen) {
-  // This is a HACK
-  var shrink = require("./shrink.js");
-  var show = require("./show.js");
-  var arbitraryBless = require("./arbitraryBless.js");
+  // Circular dependency :(
+  var string = require("./string.js");
 
-  var arb = require("./dict.js").dict(arbitraryBless({ generator: gen, shrink: shrink.noop, show: show.def }));
-  var result = arb.generator;
+  var pairGen = generatePair(string.string.generator, gen);
+  var arrayGen = generateArray(pairGen);
+  var result = arrayGen.map(utils.pairArrayToDict);
+
   return utils.curried2(result, arguments);
 }
