+++ lib/utils.js
@@ -117,6 +117,22 @@ function stringToCharArray(str) {
   return str.split("");
 }
 
+function pairArrayToDict(arrayOfPairs) {
+  var res = {};
+  arrayOfPairs.forEach(function (p) {
+    res[p[0]] = p[1];
+  });
+  return res;
+}
+
+function dictToPairArray(m) {
+  var res = [];
+  Object.keys(m).forEach(function (k) {
+    res.push([k, m[k]]);
+  });
+  return res;
+}
+
 module.exports = {
   isArray: isArray,
   isObject: isObject,
@@ -130,4 +146,6 @@ module.exports = {
   curried3: curried3,
   charArrayToString: charArrayToString,
   stringToCharArray: stringToCharArray,
+  pairArrayToDict: pairArrayToDict,
+  dictToPairArray: dictToPairArray,
 };
